requirements.txt
quamash
asynctest
networkx
attrs
duniter-mirage
duniterpy>=0.40
pytest
pytest-asyncio<0.6
pyyaml
aiohttp
async_timeout